requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
gunicorn==21.2.0
azure-cognitiveservices-speech==1.35.0
pyttsx3==2.98
comtypes==1.4.10
pypiwin32
pywin32==306 